nestSocialNetwork/backend/minorFixes
WHAT: minor fixes

WHY: Minor fixes to clean up the code

IMPACT: Low
