+++ src/moderators/moderator.interface.ts
@@ -1,0 +1,7 @@
+export interface Moderator {
+  firstName: string;
+  lastName: string;
+  username: string;
+  email: string;
+  password: string;
+}
