+++ src/moderators/moderators.model.ts
@@ -22,11 +22,3 @@ export const ModeratorSchema = new mongoose.Schema({
     required: true,
   },
 });
-
-export interface Moderator {
-  firstName: string;
-  lastName: string;
-  username: string;
-  email: string;
-  password: string;
-}
