+++ src/posts/post.interface.ts
@@ -1,0 +1,7 @@
+export interface Post {
+  title: string;
+  content: string;
+  dateCreated: string;
+  lastUpdated: string;
+  createdBy: string;
+}
