+++ src/posts/posts.service.ts
@@ -2,8 +2,8 @@ import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 
-import { Post } from './posts.model';
-import { User } from '../users/users.model';
+import { Post } from './post.interface';
+import { User } from '../users/user.interface';
 import { SocketsGateway } from 'src/sockets/sockets.gateway';
 
 @Injectable()
@@ -36,7 +36,10 @@ export class PostsService {
       const postsCount = await this.postModel.find({ createdBy: id }).count();
 
       if (pageNumber * this.PER_PAGE_ITEMS >= postsCount + this.PER_PAGE_ITEMS)
-        return { error: 'Page does not exist' };
+        throw new HttpException(
+          'Requested page does not exist',
+          HttpStatus.NOT_FOUND,
+        );
 
       const posts = await this.postModel
         .find({ createdBy: id })
@@ -45,13 +48,7 @@ export class PostsService {
         .limit(this.PER_PAGE_ITEMS);
 
       if (!posts)
-        throw new HttpException(
-          {
-            status: HttpStatus.NOT_FOUND,
-            error: 'Posts not found',
-          },
-          HttpStatus.NOT_FOUND,
-        );
+        throw new HttpException('No posts found', HttpStatus.NOT_FOUND);
 
       const mappedPosts = posts.map((post) => ({
         _id: post.id,
@@ -116,11 +113,7 @@ export class PostsService {
     const user = await this.userModel.findOne({ _id: userId });
     if (user.type === 'unpaid')
       throw new HttpException(
-        {
-          status: HttpStatus.UNAUTHORIZED,
-          error:
-            'Buy the subscription to view the feed. Go to {DOMAIN_NAME}/checkout',
-        },
+        'Buy the subscription to view the feed. Go to {DOMAIN_NAME}/checkout',
         HttpStatus.UNAUTHORIZED,
       );
 
@@ -131,7 +124,10 @@ export class PostsService {
       .count();
 
     if (pageNumber * this.PER_PAGE_ITEMS >= postsCount + this.PER_PAGE_ITEMS)
-      return { error: 'Page does not exist' };
+      throw new HttpException(
+        'Requested page does not exist',
+        HttpStatus.NOT_FOUND,
+      );
 
     const posts = await this.postModel
       .find({ createdBy: { $in: following } })
@@ -139,14 +135,7 @@ export class PostsService {
       .skip((pageNumber - 1) * this.PER_PAGE_ITEMS)
       .limit(this.PER_PAGE_ITEMS);
 
-    if (!posts)
-      throw new HttpException(
-        {
-          status: HttpStatus.NOT_FOUND,
-          error: 'Posts not found',
-        },
-        HttpStatus.NOT_FOUND,
-      );
+    if (!posts) throw new HttpException('No posts found', HttpStatus.NOT_FOUND);
 
     return {
       success: true,
@@ -175,20 +164,11 @@ export class PostsService {
       const post = await this.postModel.findOne({ _id: postId });
 
       if (!post)
-        throw new HttpException(
-          {
-            status: HttpStatus.NOT_FOUND,
-            error: 'Posts not found',
-          },
-          HttpStatus.NOT_FOUND,
-        );
+        throw new HttpException('No posts found', HttpStatus.NOT_FOUND);
 
       if (post.createdBy !== userId)
         throw new HttpException(
-          {
-            status: HttpStatus.UNAUTHORIZED,
-            error: "You cannot update another user's post",
-          },
+          "You cannot update another user's post",
           HttpStatus.UNAUTHORIZED,
         );
 
@@ -219,20 +199,11 @@ export class PostsService {
     try {
       const post = await this.postModel.findOneAndDelete({ _id: postId });
       if (!post)
-        throw new HttpException(
-          {
-            status: HttpStatus.NOT_FOUND,
-            error: 'Posts not found',
-          },
-          HttpStatus.NOT_FOUND,
-        );
+        throw new HttpException('No posts found', HttpStatus.NOT_FOUND);
 
       if (post.createdBy !== userId)
         throw new HttpException(
-          {
-            status: HttpStatus.UNAUTHORIZED,
-            error: "You cannot delete another user's post",
-          },
+          "You cannot delete another user's post",
           HttpStatus.UNAUTHORIZED,
         );
