+++ src/users/user.interface.ts
@@ -1,0 +1,10 @@
+export interface User {
+  firstName: string;
+  lastName: string;
+  username: string;
+  email: string;
+  password: string;
+  followers: string[];
+  following: string[];
+  type: string;
+}
