+++ src/users/users.model.ts
@@ -32,14 +32,3 @@ export const UserSchema = new mongoose.Schema({
     default: 'unpaid',
   },
 });
-
-export interface User {
-  firstName: string;
-  lastName: string;
-  username: string;
-  email: string;
-  password: string;
-  followers: string[];
-  following: string[];
-  type: string;
-}
